Add mypy for type checking and update Python version requirements
- Updated the supported Python version in `CONTRIBUTING.md`, `README.md`, and `pyproject.toml` to require Python 3.10 or higher.
- Added mypy to the pre-commit configuration for type checking, including additional dependencies for type hints.
- Refactored type hints across various modules to improve type safety and clarity.
- Enhanced docstrings and type annotations for better documentation and consistency.

+++ oligo_designer_toolsuite/database/_reference_database.py
@@ -5,7 +5,7 @@
 import os
 import shutil
 from pathlib import Path
-from typing import List, Union, get_args
+from typing import Any, get_args
 
 from oligo_designer_toolsuite._constants import _TYPES_REF
 from oligo_designer_toolsuite._exceptions import DatabaseError
@@ -32,20 +32,20 @@ class ReferenceDatabase:
         self.database_name = database_name
         self.dir_output = os.path.abspath(dir_output)
 
-        self.database_file = None
-        self.database_type = None
+        self.database_file: str | None = None
+        self.database_type: str | None = None
 
         self.fasta_parser = FastaParser()
         self.vcf_parser = VCFParser()
 
     def __del__(self):
         """Destructor for the ReferenceDatabase class."""
-        if os.path.exists(self.database_file):
+        if self.database_file is not None and os.path.exists(self.database_file):
             os.remove(self.database_file)
 
     def load_database_from_file(
         self,
-        files: Union[str, List[str]],
+        files: str | list[str],
         file_type: _TYPES_REF,
         database_overwrite: bool,
     ) -> None:
@@ -68,7 +68,7 @@ class ReferenceDatabase:
         AGTTGACAGACCCCAGATTAAAGTGTGTCGCGCAACAC
 
         :param files: Path(s) to the file(s) containing the database sequences.
-        :type files: Union[str, List[str]]
+        :type files: str | list[str]
         :param file_type: Type of the reference sequences (must be a valid type).
         :type file_type: _TYPES_REF["fasta", "vcf"]
         :param database_overwrite: If True, the existing database content will be cleared before loading the new sequences.
@@ -93,7 +93,7 @@ class ReferenceDatabase:
             files_in = files
             self.database_type = file_type
         elif self.database_file is not None and self.database_type == file_type:
-            files_in = files.extend(self.database_file)
+            files_in = files + [self.database_file]
         else:
             raise DatabaseError(
                 f"Cannot mix {file_type} and {self.database_type} databases. "
@@ -112,16 +112,16 @@ class ReferenceDatabase:
             self.database_file = os.path.join(self.dir_output, f"tmp_{self.database_name}.vcf.gz")
             self.vcf_parser.merge_vcf_files(files_in=files_in, file_out=self.database_file)
         else:
-            ValueError(f"Database type {self.database_type} not supported.")
+            raise DatabaseError(f"Database type {self.database_type} not supported.")
 
-    def write_database_to_file(self, filename: str, dir_output: str = None) -> str:
+    def write_database_to_file(self, filename: str, dir_output: str | None = None) -> str:
         """
         Write the loaded database to a file based on its type.
 
         :param filename: Name of the output file (without extension).
         :type filename: str
         :param dir_output: Directory path where output files will be saved. If None, the directory defined in the init function will be used.
-        :type dir_output: str
+        :type dir_output: str | None
         :return: Path to the written database file.
         :rtype: str
         :raises ValueError: If the database type is not supported or if the database is empty.
@@ -139,12 +139,12 @@ class ReferenceDatabase:
         else:
             raise DatabaseError("Database is empty. Nothing to be written to file.")
 
-    def filter_database_by_region(self, region_ids: Union[str, List[str]], keep_region: bool) -> None:
+    def filter_database_by_region(self, region_ids: str | list[str], keep_region: bool) -> None:
         """
         Filter the database to retain or remove specific regions.
 
-        :param region_ids: Region identifier(s) to process. Can be a single region ID (str) or a list of region IDs (List[str]). If None, all regions in the database are processed.
-        :type region_ids: Union[str, List[str]]
+        :param region_ids: Region identifier(s) to process. Can be a single region ID (str) or a list of region IDs (list[str]). If None, all regions in the database are processed.
+        :type region_ids: str | list[str]
         :param keep_region: Whether to keep (True) or remove (False) the specified regions.
         :type keep_region: bool
         :return: Path to the filtered database file.
@@ -171,16 +171,16 @@ class ReferenceDatabase:
     def filter_database_by_property_category(
         self,
         property_name: str,
-        property_category: Union[str, List[str]],
+        property_category: str | list[str],
         keep_if_equals_category: bool,
-    ) -> None:
+    ) -> Any:
         """
         Filter the database to retain or remove specific property categories.
 
         :param property_name: The property used for filtering.
         :type property_name: str
-        :param property_category: List of property values to filter by.
-        :type property_category: Union[str, List[str]]
+        :param property_category: list of property values to filter by.
+        :type property_category: str | list[str]
         :param keep_if_equals_category: Whether to keep (True) or remove (False) records with the specified property values.
         :type keep_if_equals_category: bool
         :return: Path to the filtered database file.
@@ -209,26 +209,31 @@ class ReferenceDatabase:
 
     def _filter_fasta_database_by_region(
         self,
-        region_ids: Union[str, List[str]],
+        region_ids: str | list[str],
         keep_region: bool,
     ):
         """
         Filter a FASTA database to retain or exclude specific regions.
         Therefore, merge fasta files and load fasta content for filtering.
 
-        :param region_ids: Region identifier(s) to process. Can be a single region ID (str) or a list of region IDs (List[str]). If None, all regions in the database are processed.
-        :type region_ids: Union[str, List[str]]
+        :param region_ids: Region identifier(s) to process. Can be a single region ID (str) or a list of region IDs (list[str]). If None, all regions in the database are processed.
+        :type region_ids: str | list[str]
         :param keep_region: Whether to keep (True) or remove (False) the specified regions.
         :type keep_region: bool
         """
+        if self.database_file:
+            file: str = self.database_file
+        else:
+            raise DatabaseError("Database file is not set. Call load_database_from_file() first.")
+
         if keep_region:
             regions_to_keep = region_ids
         else:
-            regions_fasta = self.fasta_parser.get_fasta_regions(file_fasta_in=self.database_file)
+            regions_fasta = self.fasta_parser.get_fasta_regions(file_fasta_in=file)
             regions_to_keep = [region for region in regions_fasta if region not in region_ids]
 
         fasta_sequences_filtered = self.fasta_parser.read_fasta_sequences(
-            file_fasta_in=self.database_file, region_ids=regions_to_keep
+            file_fasta_in=file, region_ids=regions_to_keep
         )
 
         file_database_filtered = os.path.join(self.dir_output, f"{self.database_name}_filtered.fna")
@@ -241,7 +246,7 @@ class ReferenceDatabase:
     def _filter_fasta_database_by_property_category(
         self,
         property_name: str,
-        property_category: Union[str, List[str]],
+        property_category: str | list[str],
         keep_if_equals_category: bool,
     ):
         """
@@ -250,12 +255,17 @@ class ReferenceDatabase:
 
         :param property_name: The property in the FASTA header used for filtering.
         :type property_name: str
-        :param property_category: List of property categories used to filter sequences.
-        :type property_category: Union[str, List[str]]
+        :param property_category: list of property categories used to filter sequences.
+        :type property_category: str | list[str]
         :param keep_if_equals_category: Whether to keep (True) or remove (False) sequences matching the property category.
         :type keep_if_equals_category: bool
         """
-        fasta_sequences = self.fasta_parser.read_fasta_sequences(file_fasta_in=self.database_file)
+        if self.database_file:
+            file: str = self.database_file
+        else:
+            raise DatabaseError("Database file is not set. Call load_database_from_file() first.")
+
+        fasta_sequences = self.fasta_parser.read_fasta_sequences(file_fasta_in=file)
 
         fasta_sequences_filtered = []
